lint with pylint and flake8
The script is checked with pylint and flake8, two defaults the
linters report as potentially harmful are edited.  For them,
the script passes the tests better, than the earlier version.
As checked by `diff -s`, the result output processing the test
data matches the previously generated results; this is interpreted
as "good enough" for now.