ci: add pre-commit hook (black and flake8)
For better consistency of the format, an automatic check (and if
necessary reformat) is provided by black.[1]  Subsequently, flake8[2]
equally can test the scripts within the git pre-commit hook; however,
issues here will not prevent a Python script to enter commit history.

For the purpose of this addition, the results of black's checks on
the Python script are not in the scope of this commit.

[1] https://github.com/psf/black
[2] https://github.com/PyCQA/flake8

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>