refactor: edit label sort to dictionary
After inspection of the clusters and determination of the number
of molecules they contain, it appears more useful to relate old
cluster labels (assigned by DW) with new ones (regardless what
sort engaged) with a dictionary instead of a list.  The aim here
is a reduction of complexity in the script.

Since the sort, and subsequent creation of the dictionary is
only about the cluster labels, rather than the clusters itself,
the function in charge was renamed accordingly.

To start this multi-step rewrite, several functions the script
are intentionally commented out.  Their reinstallment will be
gradually.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>