refactor: cluster are renamed by list approach
The actual reassignment of the clusters now is reorganized to
use old (DW assigned) labels as a dictionary key to fetch the
new labels from (as values) from said dictionary.  This is
less nested (saves two levels of indentation) and easier to
read and maintain.

At this commit, the script still is not yet fully functional
again.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>