test: reorganize pytests as `blackbox`, or `imported`
For an easier discern, pytest checks are reorganized in groups.
If they run the script from the outside, like a black box, they
are marked by `blackbox`.  Tests based on the import of one of
the script's functions are marked as `imported`.  The check for
the script's presence doesn't fit into either set.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>